# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /signup/dupcheck:
    x-swagger-router-controller: authenticate
    post:
      description: Check if the given username is already exist
      operationId: usernameDupCheck
      parameters:
        - in: body
          name: username
          description: username to check if it is duplicate
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        "201":
          description: Duplicate
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
...
  /game/recorded/{game_id}:
    x-swagger-router-controller: game
    get:
      description: Get recorded game data by game_id
      operationId: getRecordedGameById
      parameters:
        - in: path
          name: game_id
          description: game_id of recorded game to get data
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GameDataResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
...
definitions:
...
  MessageResponse:
    properties:
      message:
        type: string
...
  GameDataResponse:
    required:
      - shoot_data
      - position_data
    properties:
      position_data:
        type: array
        items:
          $ref: "#/definitions/PositionDataItem"
      shoot_data:
        type: array
        items:
          $ref: "#/definitions/ShootDataItem"
...
  PositionDataItem:
    required:
      - datetime
      - positions
      - own_team
    properties:
      datetime:
        type: string
        format: date-time
      positions:
        type: array
        items:
          $ref: "#/definitions/PositionItem"
      own_team:
        type: boolean
  ShootDataItem:
    required:
      - datetime
      - position
      - is_goal
    properties:
      datetime:
        type: string
        format: date-time
      position:
        $ref: "#/definitions/PositionItem"
      is_goal:
        type: boolean
  PositionItem:
    required:
      - px
      - py
      - team
    properties:
      px:
        type: number
        format: float
      py:
        type: number
        format: float
      team:
        type: integer
        minimum: 0
        maximum: 2
securityDefinitions:
  api_key:
    type: apiKey
    in: query
    name: api_key
security:
  - api_key: []